[TASK] Git command must happen in the collection directory
This change adapt createPullRequestFromGerritCommand method to set the
collection path as working directory and add a new error message, if the
working directory does not exist. This error message help user with wrong
remotes setup, because after the patch set branch is create it's possible
that the package directory does not exist any more.